pyproject.toml
[tool.poetry]
name = "software-inferno"
version = "0.1.0"
description = ""
authors = ["chrxn1c <rivalofficialentertainment@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
apscheduler = "^3.10.4"
gunicorn = "^22.0.0"
django-dotenv = "^1.4.2"
django-cors-headers = "^4.3.1"
setuptools = "^70.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
